fix: randomise options and fetch them from Cosmic
